backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
beautifulsoup4==4.12.2
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0
langchain-core==0.1.10
langchain-community==0.0.12
langchain-google-genai==0.0.6
python-dotenv==1.0.0
psycopg2-binary==2.9.9
pymysql==1.1.0
python-multipart==0.0.6
openai
